% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activation.R
\docType{data}
\name{Activation}
\alias{Activation}
\alias{Linear}
\alias{Sigmoid}
\alias{ReLU}
\title{Activation Function}
\format{
An object of class \code{Activation} (inherits from \code{R6}) of length 5.

An object of class \code{Activation} (inherits from \code{R6}) of length 5.

An object of class \code{Activation} (inherits from \code{R6}) of length 5.
}
\usage{
Linear

Sigmoid

ReLU
}
\description{
An activation function to specify for a neural network \link{Layer}.
}
\examples{
model <- Sequential$new(list(
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model
model$layers
model2 <- Sequential$new(list(
  Input$new(40),
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model2
model2$layers

}
\keyword{datasets}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fun}}{Activation function.}

\item{\code{der}}{Derivative function of the activation function.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Activation-new}{\code{Activation$new()}}
\item \href{#method-Activation-clone}{\code{Activation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Activation-new"></a>}}
\if{latex}{\out{\hypertarget{method-Activation-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Activation$new(fun, der)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{Activation function.}

\item{\code{der}}{Derivative function of the activation function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Activation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Activation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Activation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
