% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dense.R
\name{Dense}
\alias{Dense}
\title{Dense Layer}
\description{
A fully (densely) connected neural network \link{Layer}.
}
\examples{

## ------------------------------------------------
## Method `Dense$new`
## ------------------------------------------------

model <- Sequential$new(list(
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model
model$layers
model2 <- Sequential$new(list(
  Input$new(40),
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model2
model2$layers
}
\section{Super class}{
\code{\link[percept:Layer]{percept::Layer}} -> \code{Dense}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{units}}{Number of nodes in the layer.}

\item{\code{activation}}{\link{Activation} function to apply to the output.}

\item{\code{weights}}{Matrix of connection weights between this layer
and the previous layer in the model.}

\item{\code{input}}{The input supplied to the layer upon feed forward.}

\item{\code{output}}{The output of the layer upon feed forward.}

\item{\code{deltas}}{Delta value of each node calculated upon backpropagation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dense-new}{\code{Dense$new()}}
\item \href{#method-Dense-create_weights}{\code{Dense$create_weights()}}
\item \href{#method-Dense-feed_forward}{\code{Dense$feed_forward()}}
\item \href{#method-Dense-clone}{\code{Dense$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dense-new"></a>}}
\if{latex}{\out{\hypertarget{method-Dense-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense$new(units, activation = Linear)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{Number of nodes in the layer.}

\item{\code{activation}}{\link{Activation} function to apply to the output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model <- Sequential$new(list(
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model
model$layers
model2 <- Sequential$new(list(
  Input$new(40),
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model2
model2$layers
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dense-create_weights"></a>}}
\if{latex}{\out{\hypertarget{method-Dense-create_weights}{}}}
\subsection{Method \code{create_weights()}}{
Randomly initialises weights based on the number of input nodes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense$create_weights(input_units)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_units}}{Number of input nodes to the layer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dense-feed_forward"></a>}}
\if{latex}{\out{\hypertarget{method-Dense-feed_forward}{}}}
\subsection{Method \code{feed_forward()}}{
Feeds an input through the layer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense$feed_forward(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Numeric array of size \verb{(batch_size, input_units)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Output of size \verb{(batch_size, output_units)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dense-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Dense-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dense$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
