% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Input.R
\name{Input}
\alias{Input}
\title{Input Layer}
\description{
An input \link{Layer} to optionally specify in a neural network model.
}
\examples{

## ------------------------------------------------
## Method `Input$new`
## ------------------------------------------------

model <- Sequential$new(list(
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model
model$layers
model2 <- Sequential$new(list(
  Input$new(40),
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model2
model2$layers
}
\section{Super class}{
\code{\link[percept:Layer]{percept::Layer}} -> \code{Input}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{units}}{Number of nodes in the layer.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Input-new}{\code{Input$new()}}
\item \href{#method-Input-clone}{\code{Input$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-new"></a>}}
\if{latex}{\out{\hypertarget{method-Input-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Input$new(units)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{units}}{Number of nodes in the layer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model <- Sequential$new(list(
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model
model$layers
model2 <- Sequential$new(list(
  Input$new(40),
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model2
model2$layers
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Input-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Input-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Input$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
