% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percept-package.R
\docType{package}
\name{percept-package}
\alias{percept}
\alias{percept-package}
\title{percept: Feed-Forward Neural Networks}
\description{
Build feed-forward neural networks with multiple hidden layers and activation functions.
}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
