% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-Sequential.R
\name{plot.Sequential}
\alias{plot.Sequential}
\title{Sequential Model Plot}
\usage{
\method{plot}{Sequential}(
  x,
  max_nodes = 30,
  main = "Sequential Model",
  colors = c("red", "white", "blue"),
  ...
)
}
\arguments{
\item{max_nodes}{Maximum number of nodes plotted for each layer.}

\item{main}{Title of the plot.}

\item{colors}{Colors of negative, zero, and positive weights respectively.}

\item{...}{Ignored.}
}
\value{
\code{NULL}.
}
\description{
Plots the structure of a \link{Sequential} model.
}
