% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequential.R
\name{Sequential}
\alias{Sequential}
\title{Sequential Model}
\description{
A neural network model for a plain stack of layers where each \link{Layer} has
one input and one output array.
}
\examples{

## ------------------------------------------------
## Method `Sequential$new`
## ------------------------------------------------

model <- Sequential$new(list(
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model
model$layers
model2 <- Sequential$new(list(
  Input$new(40),
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model2
model2$layers
}
\seealso{
\code{\link[=to_categorical]{to_categorical()}}, \code{\link[=train_split]{train_split()}}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{layers}}{List of the \link{Layer}s of the model.}

\item{\code{input_units}}{Number of input nodes to the model.}

\item{\code{learning_rate}}{Learning rate used to fit the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sequential-new}{\code{Sequential$new()}}
\item \href{#method-Sequential-add}{\code{Sequential$add()}}
\item \href{#method-Sequential-pop}{\code{Sequential$pop()}}
\item \href{#method-Sequential-feed_forward}{\code{Sequential$feed_forward()}}
\item \href{#method-Sequential-backpropagate}{\code{Sequential$backpropagate()}}
\item \href{#method-Sequential-fit}{\code{Sequential$fit()}}
\item \href{#method-Sequential-clone}{\code{Sequential$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sequential-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sequential-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
Calls the \code{add()} method on each \link{Layer} in \code{layers}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$new(layers = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layers}}{List of \link{Layer}s to add to the model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{model <- Sequential$new(list(
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model
model$layers
model2 <- Sequential$new(list(
  Input$new(40),
  Dense$new(20, ReLU),
  Dense$new(10, Sigmoid)
))
model2
model2$layers
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sequential-add"></a>}}
\if{latex}{\out{\hypertarget{method-Sequential-add}{}}}
\subsection{Method \code{add()}}{
Adds a \link{Layer} to the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$add(layer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layer}}{\link{Layer}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sequential-pop"></a>}}
\if{latex}{\out{\hypertarget{method-Sequential-pop}{}}}
\subsection{Method \code{pop()}}{
Removes a \link{Layer} from the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$pop()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\link{Layer} removed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sequential-feed_forward"></a>}}
\if{latex}{\out{\hypertarget{method-Sequential-feed_forward}{}}}
\subsection{Method \code{feed_forward()}}{
Feeds an input forward in the model.
Calls the \code{feed_forward()} method of each \link{Layer} in the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$feed_forward(input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Numeric array of size \verb{(batch_size, input_units)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Output of the last \link{Layer} in the model, of size \verb{(batch_size, output_units)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sequential-backpropagate"></a>}}
\if{latex}{\out{\hypertarget{method-Sequential-backpropagate}{}}}
\subsection{Method \code{backpropagate()}}{
Backpropagates to update the weights of each \link{Layer}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$backpropagate(expected)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expected}}{Numeric array of size \verb{(batch_size, output_units)}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sequential-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Sequential-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the model to inputs and expected outputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$fit(
  input,
  expected,
  batch_size,
  epochs,
  learning_rate,
  verbose = epochs\%/\%10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{Numeric array of size \verb{(>batch_size, input_units)}.}

\item{\code{expected}}{Numeric array of size \verb{(>batch_size, output_units)}.}

\item{\code{batch_size}}{Batch size for each \code{feed_forward()} and
\code{backpropagate()} call.}

\item{\code{epochs}}{Number of times a batch is used to update the
weights of the model.}

\item{\code{learning_rate}}{Rate at which weights are adjusted.}

\item{\code{verbose}}{How often to print the epoch number.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sequential-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sequential-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sequential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
