% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_categorical.R
\name{to_categorical}
\alias{to_categorical}
\alias{to_factor}
\title{Categorical Data}
\usage{
to_categorical(f)

to_factor(x, labels)
}
\arguments{
\item{f}{Factor.}

\item{x}{Numeric matrix where each row is an observation.}

\item{labels}{Passed to \code{\link[=factor]{factor()}}.}
}
\value{
For \code{to_categorical()}, numeric matrix. For \code{to_factor()}, factor.
}
\description{
Convert between factor and model input/output batch.
}
\seealso{
\link{Sequential}.
}
