% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_split.R
\name{train_split}
\alias{train_split}
\title{Train and Test Data}
\usage{
train_split(x, train = 0.8)
}
\arguments{
\item{x}{Data frame.}

\item{train}{Proportion of rows to allocate to the training set.}
}
\value{
Logical vector indicating if each row is in the training set.
}
\description{
Randomly split rows of a data frame into train and test sets.
}
\seealso{
\code{\link[=to_categorical]{to_categorical()}}, \link{Sequential}.
}
