% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percept-package.R
\docType{package}
\name{percept-package}
\alias{percept}
\alias{percept-package}
\title{percept: Feed-Forward Neural Networks}
\description{
Build feed-forward neural networks with multiple hidden layers and activation functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrryt/percept}
  \item Report bugs at \url{https://github.com/hrryt/percept/issues}
}

}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
